+++ src/components/ContextMenu/style/index.ts
@@ -3,9 +3,9 @@ import type { AliasToken, GenerateStyle } from 'antd/es/theme/internal';
 import type { CSSObject } from '@ant-design/cssinjs';
 import type { FullToken } from '@ant-design/cssinjs-utils';
 
-type OverflowTagsToken = FullToken<{ ''?: object }, AliasToken, ''>;
+type ContextMenuToken = FullToken<{ ''?: object }, AliasToken, ''>;
 
-const genStyle: GenerateStyle<OverflowTagsToken> = (token): CSSObject => {
+const genStyle: GenerateStyle<ContextMenuToken> = (token): CSSObject => {
   const { componentCls } = token;
   return {
     [componentCls]: {