+++ src/components/EditableText/index.tsx
@@ -1,6 +1,6 @@
 import type { CSSProperties, ReactNode } from 'react';
 import React, { useContext, useEffect, useMemo, useState } from 'react';
-import { ConfigProvider, Flex, Typography } from 'antd';
+import { Flex, Typography } from 'antd';
 import type { EllipsisConfig } from 'antd/es/typography/Base';
 import type { LinkProps } from 'antd/es/typography/Link';
 import type { ParagraphProps } from 'antd/es/typography/Paragraph';
@@ -9,6 +9,7 @@ import type { TitleProps } from 'antd/es/typography/Title';
 import names from 'classnames';
 import { EditOutlined } from '@ant-design/icons';
 import useT from '../../hooks/useT';
+import ConfigProvider from '../ConfigProvider';
 import EditableTextForm, { type EditableFormProps, type RenderInputInterface } from './form';
 import useStyle from './style';
 
@@ -219,7 +220,7 @@ const EditableText = <
   } = props;
 
   const { getPrefixCls } = useContext(ConfigProvider.ConfigContext);
-  const prefixCls = getPrefixCls('easy-editable-text', prefixClsInProps);
+  const prefixCls = getPrefixCls('editable-text', prefixClsInProps);
   const [wrapCSSVar, hashId, cssVarCls] = useStyle(prefixCls);
   const t = useT();
   const [isEditing, setIsEditing] = useState<boolean>(editing);